Adds the Employee class
The Employee is responsible for running one job at a time. Each job it
runs is executed in its own forked process. If an exception is raised
from within the job this results in exit code 1 which makes the Employee
report the job as failed, otherwise the exist code is 0 and the Employee
reports the job as completed.